separate route and functions
