+++ routes/betRoutes.js
@@ -3,13 +3,13 @@ const {
   getBetsByGameID, 
   postBets, 
   putBets, 
-  deleteBets } = require('./betRoutesHandlers');
+  deleteBets 
+} = require('./betRoutesHandlers');
 
 module.exports = (app) => {
   // Get
   app.get('/api/bets', getBets);
   app.get('/api/bets/:game_id', getBetsByGameID);
-
   // Create
   app.post('/api/bets', postBets);
